+++ CHANGES.md
@@ -5,6 +5,8 @@
 
 - Allow to config/set `addressing_style` for S3 object storage. (default='virtual')
 
+- Added `LRUCache` and `SyncLRUCache`.
+
 
 ## 0.16.2 (2024-07-16)
 ----------------------